Major change: switch ZTimestamp from second to nanosecond precision.
Part of the design goal of ZTimestamp was to improve efficiency and simplicity. A compact representation was part of that: just 2 small integers. That was in the 32-bit era. We are now used to 64-bit since many years. This makes it possible to store nanosecondsSinceMidnight in a small integer.

The internal fields of ZTimestamp are now jnd and ns.

Adapt all necessary methods. Make sure all tests pass.

ZTimestamp>>#rounded converts the new, more precise instances to the older second precision variant.